refactor: 함수 및 훅 분리
